requirements.txt

Flask
xgboost
scikit-learn
pandas
numpy
python-whois
requests
